Rename openlr_simple_parser.* to openlr_model_xml.*
