+++ openlr/CMakeLists.txt
@@ -4,12 +4,12 @@ set(
   SRC
   openlr_model.cpp
   openlr_model.hpp
+  openlr_model_xml.cpp
+  openlr_model_xml.hpp
   openlr_sample.cpp
   openlr_sample.hpp
   openlr_simple_decoder.cpp
   openlr_simple_decoder.hpp
-  openlr_simple_parser.cpp
-  openlr_simple_parser.hpp
   road_info_getter.cpp
   road_info_getter.hpp
   road_type_checkers.cpp
